# inventory-api/src/main/resources/openapi.yaml
{"openapi":"3.0.1","info":{"title":"Inventory & Order API","description":"API for managing distribution centers, inventory, orders, and products.","contact":{"name":"Robert Eftene","email":"roberteftene25@gmail.com"},"version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Products","description":"Endpoints related to product management"},{"name":"Orders","description":"Operations related to orders"},{"name":"Inventory","description":"Operations related to inventory items"}],"paths":{"/api/products":{"post":{"tags":["Products"],"summary":"Create a new product","description":"Adds a new product to the catalog and returns the created product.","operationId":"createProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductDto"}}},"required":true},"responses":{"201":{"description":"Product created successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductDto"}}}},"400":{"description":"Invalid product data"}}}},"/api/orders":{"get":{"tags":["Orders"],"summary":"Get all orders","description":"Returns all orders with their associated order items","operationId":"getOrders","responses":{"200":{"description":"List of orders","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrderDto"}}}}},"403":{"description":"Unauthorized or insufficient permissions","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrderDto"}}}}}}},"post":{"tags":["Orders"],"summary":"Create a new order","description":"Creates a new order with its order items","operationId":"createOrder","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrderDto"}}},"required":true},"responses":{"200":{"description":"Order created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderDto"}}}},"400":{"description":"Invalid request body","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderCreationResponse"}}}},"403":{"description":"Unauthorized access","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderCreationResponse"}}}}}}},"/api/inventory":{"get":{"tags":["Inventory"],"summary":"Get inventory items by distribution center and product SKU","description":"Returns all inventory entries for a given distribution center and product SKU. Can be used to check stock levels and expiration dates.","operationId":"getInventory","parameters":[{"name":"dcId","in":"query","description":"ID of the distribution center","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"sku","in":"query","description":"SKU of the product","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"List of inventory items","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/InventoryDto"}}}}},"400":{"description":"Missing or invalid parameters","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/InventoryDto"}}}}},"404":{"description":"No inventory found for given parameters","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/InventoryDto"}}}}}}}}},"components":{"schemas":{"ProductDto":{"required":["description","name","price","sku"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"sku":{"type":"string"},"description":{"type":"string"},"price":{"type":"number","format":"double"},"category":{"type":"string","enum":["DAIRY","PRODUCE","BEVERAGES","CLEANING","PERSONAL_CARE","HEALTH","ELECTRONICS","TOYS","FURNITURE"]}}},"OrderDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"requestedAt":{"type":"string","format":"date-time"},"orderStatus":{"type":"string","enum":["PENDING","PROCESSING","PARTIALLY_FULFILLED","FULFILLED","CANCELLED","FAILED","ON_HOLD"]},"orderItems":{"type":"array","items":{"$ref":"#/components/schemas/OrderItemDto"}}}},"OrderItemDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"productId":{"type":"integer","format":"int64"},"quantityRequested":{"type":"integer","format":"int32"},"productName":{"type":"string"}}},"Address":{"type":"object","properties":{"street":{"type":"string"},"city":{"type":"string"},"zipCode":{"type":"string"}}},"CreateOrderDto":{"required":["billingAddress","orderItems","shippingAddress"],"type":"object","properties":{"billingAddress":{"$ref":"#/components/schemas/Address"},"shippingAddress":{"$ref":"#/components/schemas/Address"},"orderItems":{"type":"array","items":{"$ref":"#/components/schemas/CreateOrderItemDto"}}}},"CreateOrderItemDto":{"required":["dcId","productId"],"type":"object","properties":{"quantityRequested":{"minimum":1,"type":"integer","format":"int32"},"productId":{"type":"integer","format":"int64"},"dcId":{"type":"integer","format":"int64"}}},"OrderCreationResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"status":{"type":"string","enum":["PENDING","PROCESSING","PARTIALLY_FULFILLED","FULFILLED","CANCELLED","FAILED","ON_HOLD"]}}},"InventoryDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"quantity":{"type":"integer","format":"int32"},"expiresAt":{"type":"string","format":"date"},"updatedAt":{"type":"string","format":"date"}}}}}}